1. Add command line control of random seed in speed testing programs. 2. Improve documentation of the choice of random seed. 3. Change log messages from command start and done to begin and end command.
command.omh: discuss output tables and log messages.

+++ bin/time_memory.sh
@@ -1,3 +1,4 @@
+# /usr/bin/bash -e
 # $Id:$
 #  --------------------------------------------------------------------------
 # dismod_at: Estimating Disease Rates as Functions of Age and Time
@@ -14,7 +15,15 @@ then
 	exit 1
 fi
 # -----------------------------------------------------------------------------
-python3 speed/simulated.py | tee build/speed/time.out
+if [ "$1" == '' ]
+then
+	echo 'usage: time_memory random_seed'
+	echo 'use zero for choosing seed from clock'
+	exit 1
+fi
+random_seed="$1"
+# -----------------------------------------------------------------------------
+python3 speed/simulated.py $random_seed | tee build/speed/time.out
 cd build/speed
 ../devel/dismod_at example.db start
 rm memory.out.*
